[orbis-kernel] Fix sys_namedobj_delete+

+++ orbis-kernel/include/orbis/sys/sysproto.hpp
@@ -645,7 +645,7 @@ SysResult sys_osem_post(Thread *thread /* TODO */);
 SysResult sys_osem_cancel(Thread *thread /* TODO */);
 SysResult sys_namedobj_create(Thread *thread, ptr<const char[32]> name,
                               ptr<void> object, uint16_t type);
-SysResult sys_namedobj_delete(Thread *thread, uint16_t id, uint16_t type);
+SysResult sys_namedobj_delete(Thread *thread, uint id, uint16_t type);
 SysResult sys_set_vm_container(Thread *thread /* TODO */);
 SysResult sys_debug_init(Thread *thread /* TODO */);
 SysResult sys_suspend_process(Thread *thread, pid_t pid);
